+++ packages/video_player_avplay/lib/video_player_platform_interface.dart
@@ -345,6 +345,11 @@ enum StreamingPropertyType {
   /// "STARTFRAGMENT=" For live content playback, defines the start fragment number.
   /// "FIXED_MAX_RESOLUTION=max_widthXmax_height". Only if the given media URI such as mpd in MPEG-DASH or m3u8 in HLS through open()
   ///  method doesn't describe entire required video resolutions,application should use this attribute to complete the resolution information for the player.
+  /// "MAX_RESOLUTION=widthXheight". Only specifies the maximum acceptable video resolution for DASH adaptive streaming.
+  /// "MAX_FRAMERATE=framerate". Only specifies the maximum acceptable video framerate for DASH adaptive streaming.
+  /// "UPDATE_SAME_LANGUAGE_CODE=language_code". Only available for DASH stream. Update the language code in manifest like lang="'en'+'i'", where "i" will be an integer
+  /// when there are more than one adaptation set with same language code. The value of language_code be like 0, 1 or others as defined in the manifest.
+  /// "OPEN_SUBTITLE_STYLE=TRUE". Only available for DASH stream. Enable subtitle style for DASH stream.
   adaptiveInfo,
 
   /// Forces the player to use the 4K UHD decoder. Its parameter can be the string "TRUE" or "FALSE".
@@ -382,34 +387,6 @@ enum StreamingPropertyType {
   /// Property to select the Scaler type, By Default MAIN Scaler selected.
   inAppMultiView,
 
-  /// Specifies the maximum acceptable video resolution for DASH adaptive streaming.
-  ///
-  /// This property allows you to set an upper limit on the video resolution
-  /// (width x height) that the DASH player can select during playback.
-  /// The player will not choose any resolution higher than the specified maximum.
-  ///
-  /// The value for this property must be a string in the format 'widthXheight',
-  /// for example, '1920X1080' to set the maximum acceptable resolution to 1080p.
-  /// The player will then select from resolutions up to and including 1080p.
-  ///
-  /// **Important**: The set maximum resolution cannot be lower than the minimum
-  /// resolution available in the stream's manifest.
-  unwantedResolution,
-
-  /// Specifies the maximum acceptable video framerate for DASH adaptive streaming.
-  ///
-  /// This property allows you to set an upper limit on the video framerate (in frames
-  /// per second) that the DASH player can select during playback. The player
-  /// will not choose any framerate higher than the specified maximum.
-  ///
-  /// The value for this property should be a string representing the numerical framerate,
-  /// for example, '30' to set the maximum acceptable framerate to 30fps. The player
-  /// will then select from framerates up to and including 30fps.
-  ///
-  /// **Important**: The set maximum framerate cannot be lower than the minimum
-  /// framerate available in the stream's manifest.
-  unwantedFramerate,
-
   /// The audio track info of the DASH stream.
   audioStreamInfo,
 
@@ -418,14 +395,6 @@ enum StreamingPropertyType {
 
   /// The video track info of the DASH stream.
   videoStreamInfo,
-
-  /// Only available for DASH stream.
-  ///
-  /// Update the language code in manifest like lang="'en'+'i'", where "i" will be an integer
-  /// when there are more than one adaptation set with same language code.
-  ///
-  /// The value for this property is an integer string: '0','1' or others as defined in the manifest.
-  updateSameLanguageCode,
 
   /// Sets the DASH authentication token to be used before the player is initialized.
   ///
@@ -449,6 +418,11 @@ enum StreamingPropertyType {
   ///
   /// Whether to enable the function of obtaining http header. 'TRUE' or others.
   openHttpHeader,
+
+  /// Only available for DASH stream.
+  ///
+  /// To control is force enable if can get manifest content callback. 'TRUE' or others.
+  openManifest,
 }
 
 /// The different types of buffer configurations that can be set on the player.
@@ -554,10 +528,10 @@ class VideoEvent {
     this.duration,
     this.size,
     this.buffered,
-    this.text,
     this.isPlaying,
-    this.subtitleAttributes,
     this.adInfo,
+    this.manifestInfo,
+    this.subtitlesInfo,
   });
 
   /// The type of the event.
@@ -578,23 +552,25 @@ class VideoEvent {
   /// Only used if [eventType] is [VideoEventType.bufferingUpdate].
   final int? buffered;
 
-  /// Subtitle text of the video.
-  ///
-  /// Only used if [eventType] is [VideoEventType.subtitleUpdate].
-  final String? text;
-
   /// Whether the video is currently playing.
   ///
   /// Only used if [eventType] is [VideoEventType.isPlayingStateUpdate].
   final bool? isPlaying;
 
-  /// Attributes of the video subtitle.
-  final List<dynamic>? subtitleAttributes;
-
   /// The ad info from dash.
   ///
   /// Only used if [eventType] is [VideoEventType.adFromDash].
   final Map<Object?, Object?>? adInfo;
+
+  /// The manifest information in dash.
+  ///
+  /// Only used if [eventType] is [VideoEventType.manifestInfoUpdated].
+  final String? manifestInfo;
+
+  /// The subtitle information of the video.
+  ///
+  /// Only used if [eventType] is [VideoEventType.subtitleUpdate].
+  final SubtitlesInfo? subtitlesInfo;
 
   @override
   bool operator ==(Object other) {
@@ -605,10 +581,10 @@ class VideoEvent {
             duration == other.duration &&
             size == other.size &&
             buffered == other.buffered &&
-            text == other.text &&
             isPlaying == other.isPlaying &&
-            subtitleAttributes == other.subtitleAttributes &&
-            adInfo == other.adInfo;
+            mapEquals(adInfo, other.adInfo) &&
+            manifestInfo == other.manifestInfo &&
+            subtitlesInfo == other.subtitlesInfo;
   }
 
   @override
@@ -617,10 +593,10 @@ class VideoEvent {
       duration.hashCode ^
       size.hashCode ^
       buffered.hashCode ^
-      text.hashCode ^
       isPlaying.hashCode ^
-      subtitleAttributes.hashCode ^
-      adInfo.hashCode;
+      adInfo.hashCode ^
+      manifestInfo.hashCode ^
+      subtitlesInfo.hashCode;
 }
 
 /// Type of the event.
@@ -657,6 +633,9 @@ enum VideoEventType {
 
   /// The ad event from dash.
   adFromDash,
+
+  /// The manifest updated in dash.
+  manifestInfoUpdated,
 
   /// An unknown event has been received.
   unknown,
@@ -725,6 +704,35 @@ class DurationRange {
 
   @override
   int get hashCode => start.hashCode ^ end.hashCode;
+}
+
+/// The subtitle information of the video. Contains the text duration, text and attributes, and picture info.
+@immutable
+class SubtitlesInfo {
+  /// Creates an instance of [SubtitlesInfo].
+  const SubtitlesInfo(this.textDuration, this.textsInfo, this.pictureInfo);
+
+  /// The duration of text.
+  final int? textDuration;
+
+  /// Text and attributes of the video subtitle.
+  final List<dynamic>? textsInfo;
+
+  /// Subtitle picture info of the video. Includes the picture and its width and height.
+  final Map<Object?, Object?>? pictureInfo;
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        other is SubtitlesInfo &&
+            listEquals(textsInfo, other.textsInfo) &&
+            textDuration == other.textDuration &&
+            mapEquals(pictureInfo, other.pictureInfo);
+  }
+
+  @override
+  int get hashCode =>
+      textDuration.hashCode ^ textsInfo.hashCode ^ pictureInfo.hashCode;
 }
 
 /// [VideoPlayerOptions] can be optionally used to set additional player settings