+++ packages/video_player_avplay/tizen/src/plus_player.cc
@@ -14,6 +14,12 @@
 #include "rapidjson/stringbuffer.h"
 #include "rapidjson/writer.h"
 
+#define STB_IMAGE_WRITE_IMPLEMENTATION
+#include "stb_image_write.h"
+
+#define SUBTITLE_ATTR_TYPE_COUNT 37
+#define SUBTITLE_DEFAULT_TEXT_LINE 1
+
 namespace video_player_avplay_tizen {
 
 static std::vector<std::string> split(const std::string &s, char delim) {
@@ -23,6 +29,23 @@ static std::vector<std::string> split(const std::string &s, char delim) {
   while (getline(ss, item, delim)) {
     tokens.push_back(item);
   }
+  return tokens;
+}
+
+static std::vector<std::string> split_by_string(const std::string &s,
+                                                const std::string &delimiter) {
+  std::vector<std::string> tokens;
+  const size_t delim_len = delimiter.length();
+  size_t start = 0;
+  size_t pos = s.find(delimiter, start);
+
+  while (pos != std::string::npos) {
+    tokens.push_back(s.substr(start, pos - start));
+    start = pos + delim_len;
+    pos = s.find(delimiter, start);
+  }
+
+  tokens.push_back(s.substr(start));
   return tokens;
 }
 
@@ -48,6 +71,9 @@ PlusPlayer::PlusPlayer(flutter::BinaryMessenger *messenger,
 
 PlusPlayer::~PlusPlayer() {
   if (player_) {
+    if (drm_manager_) {
+      drm_manager_->StopDrmSession();
+    }
     Stop(player_);
     Close(player_);
     UnregisterListener(player_);
@@ -226,6 +252,9 @@ bool PlusPlayer::Activate() {
 
 bool PlusPlayer::Deactivate() {
   if (is_prebuffer_mode_) {
+    if (drm_manager_) {
+      drm_manager_->StopDrmSession();
+    }
     Stop(player_);
     return true;
   }
@@ -411,7 +440,7 @@ bool PlusPlayer::SetDisplay() {
     return false;
   }
   bool ret = ::SetDisplay(player_, plusplayer::DisplayType::kOverlay,
-                          resource_id, x, y, width, height);
+                          resource_id, 0, 0, width, height);
   if (!ret) {
     LOG_ERROR("[PlusPlayer] Player fail to set display.");
     return false;
@@ -532,21 +561,21 @@ flutter::EncodableList PlusPlayer::GetTrackInfo(std::string track_type) {
 
   flutter::EncodableList trackSelections = {};
   if (type == plusplayer::TrackType::kTrackTypeVideo) {
-    LOG_INFO("[PlusPlayer] Video track count: %d", track_count);
+    LOG_DEBUG("[PlusPlayer] Video track count: %d", track_count);
     for (const auto &track : track_info) {
       if (track.type == plusplayer::kTrackTypeVideo) {
         trackSelections.push_back(ParseVideoTrack(track));
       }
     }
   } else if (type == plusplayer::TrackType::kTrackTypeAudio) {
-    LOG_INFO("[PlusPlayer] Audio track count: %d", track_count);
+    LOG_DEBUG("[PlusPlayer] Audio track count: %d", track_count);
     for (const auto &track : track_info) {
       if (track.type == plusplayer::kTrackTypeAudio) {
         trackSelections.push_back(ParseAudioTrack(track));
       }
     }
   } else if (type == plusplayer::TrackType::kTrackTypeSubtitle) {
-    LOG_INFO("[PlusPlayer] Subtitle track count: %d", track_count);
+    LOG_DEBUG("[PlusPlayer] Subtitle track count: %d", track_count);
     for (const auto &track : track_info) {
       if (track.type == plusplayer::kTrackTypeSubtitle) {
         trackSelections.push_back(
@@ -615,6 +644,12 @@ bool PlusPlayer::SetTrackSelection(int32_t track_id, std::string track_type) {
 bool PlusPlayer::SetDrm(const std::string &uri, int drm_type,
                         const std::string &license_server_url) {
   drm_manager_ = std::make_unique<DrmManager>();
+  DrmManager::ErrorCallback drm_error_callback =
+      [this](const std::string &error_code, const std::string &error_message) {
+        this->SendError(error_code, error_message);
+      };
+  drm_manager_->SetErrorCallback(drm_error_callback);
+
   if (!drm_manager_->CreateDrmSession(drm_type, true)) {
     LOG_ERROR("[PlusPlayer] Fail to create drm session.");
     return false;
@@ -708,17 +743,24 @@ void PlusPlayer::SetStreamingProperty(const std::string &type,
   if ((!create_message_.format_hint() ||
        create_message_.format_hint()->empty() ||
        *create_message_.format_hint() != "dash") &&
-      (type == "OPEN_HTTP_HEADER" || type == "TOKEN" ||
-       type == "UNWANTED_FRAMERATE" || type == "UNWANTED_RESOLUTION" ||
-       type == "UPDATE_SAME_LANGUAGE_CODE")) {
+      (type == "OPEN_HTTP_HEADER" || type == "TOKEN")) {
     LOG_ERROR("[PlusPlayer] Only support streaming property type: %s for DASH!",
               type.c_str());
     return;
   }
 
-  LOG_INFO("[PlusPlayer] SetStreamingProp: type[%s], value[%s]", type.c_str(),
-           value.c_str());
-  ::SetStreamingProperty(player_, type, value);
+  if (type == "ADAPTIVE_INFO") {
+    std::vector<std::string> properties = split(value, ';');
+    for (const auto &property : properties) {
+      LOG_INFO("[PlusPlayer] SetStreamingProp: type[%s], value[%s]",
+               type.c_str(), property.c_str());
+      ::SetStreamingProperty(player_, type, property);
+    }
+  } else {
+    LOG_INFO("[PlusPlayer] SetStreamingProp: type[%s], value[%s]", type.c_str(),
+             value.c_str());
+    ::SetStreamingProperty(player_, type, value);
+  }
 }
 
 bool PlusPlayer::SetDisplayRotate(int64_t rotation) {
@@ -766,6 +808,10 @@ bool PlusPlayer::StopAndClose() {
   if (player_state < plusplayer::State::kReady) {
     LOG_INFO("[PlusPlayer] Player already stop, nothing to do.");
     return true;
+  }
+
+  if (drm_manager_) {
+    drm_manager_->StopDrmSession();
   }
 
   if (!::Stop(player_)) {
@@ -1085,11 +1131,9 @@ void PlusPlayer::OnPrepareDone(bool ret, void *user_data) {
   if (!SetDisplayVisible(self->player_, true)) {
     LOG_ERROR("[PlusPlayer] Fail to set display visible.");
   }
-
   if (!self->is_initialized_ && ret) {
     self->SendInitialized();
   }
-
   if (self->is_restored_ && ret) {
     self->SendRestored();
   }
@@ -1132,15 +1176,40 @@ void PlusPlayer::OnSubtitleData(char *data, const int size,
                                 const uint64_t duration,
                                 plusplayer::SubtitleAttributeListPtr attr_list,
                                 void *user_data) {
-  LOG_INFO("[PlusPlayer] Subtitle updated, duration: %llu, text: %s", duration,
-           data);
+  LOG_INFO("[PlusPlayer] Subtitle updated, duration: %llu, type: %d", duration,
+           type);
+
+  if (!data || type == plusplayer::SubtitleType::kInvalid) {
+    LOG_ERROR("[PlusPlayer] Subtitle type is invalid or data is null.");
+    return;
+  }
+
   PlusPlayer *self = reinterpret_cast<PlusPlayer *>(user_data);
 
+  int text_lines_count = SUBTITLE_DEFAULT_TEXT_LINE;
+  std::vector<std::string> text_lines;
+  if (type != plusplayer::SubtitleType::kPicture) {
+    std::string text_data(data);
+    text_lines = split_by_string(text_data, "+++");
+    text_lines_count = text_lines.size();
+  }
+
+  std::vector<int> line_attr_index(SUBTITLE_ATTR_TYPE_COUNT,
+                                   text_lines_count - 1);
+
+  // Pre-parse attributes to find width/height for picture subtitles.
+  double picture_width = 0.0;
+  double picture_height = 0.0;
+
   plusplayer::SubtitleAttributeList *attrs = attr_list.get();
-  flutter::EncodableList attributes_list;
+
+  std::vector<flutter::EncodableList> attributes_lines(
+      text_lines_count, flutter::EncodableList());
+
   for (auto attr = attrs->begin(); attr != attrs->end(); attr++) {
-    LOG_INFO("[PlusPlayer] Subtitle update: type: %d, start: %u, end: %u",
-             attr->type, attr->start_time, attr->stop_time);
+    LOG_INFO(
+        "[PlusPlayer] SubtitleAttr update: attrType: %d, start: %u, end: %u.",
+        attr->type, attr->start_time, attr->stop_time);
     flutter::EncodableMap attributes = {
         {flutter::EncodableValue("attrType"),
          flutter::EncodableValue(attr->type)},
@@ -1149,6 +1218,11 @@ void PlusPlayer::OnSubtitleData(char *data, const int size,
         {flutter::EncodableValue("stopTime"),
          flutter::EncodableValue((int64_t)attr->stop_time)},
     };
+
+    if (attr->value == nullptr) {
+      LOG_ERROR("[PlusPlayer] SubtitleAttr value is null");
+      return;
+    }
 
     switch (attr->type) {
       case plusplayer::kSubAttrRegionXPos:
@@ -1164,23 +1238,27 @@ void PlusPlayer::OnSubtitleData(char *data, const int size,
       case plusplayer::kSubAttrWebvttCueLine:
       case plusplayer::kSubAttrWebvttCueSize:
       case plusplayer::kSubAttrWebvttCuePosition: {
-        intptr_t value_temp = reinterpret_cast<intptr_t>(attr->value);
-        float value_float;
-        std::memcpy(&value_float, &value_temp, sizeof(float));
-        LOG_INFO("[PlusPlayer] Subtitle update: value<float>: %f", value_float);
+        const float *value_float = static_cast<const float *>(attr->value);
+        LOG_INFO("[PlusPlayer] Subtitle update: value<float>: %f",
+                 *value_float);
         attributes[flutter::EncodableValue("attrValue")] =
-            flutter::EncodableValue((double)value_float);
+            flutter::EncodableValue((double)*value_float);
+
+        if (attr->type == plusplayer::kSubAttrRegionWidth &&
+            type == plusplayer::SubtitleType::kPicture) {
+          picture_width = (double)*value_float;
+        }
+        if (attr->type == plusplayer::kSubAttrRegionHeight &&
+            type == plusplayer::SubtitleType::kPicture) {
+          picture_height = (double)*value_float;
+        }
       } break;
       case plusplayer::kSubAttrWindowLeftMargin:
       case plusplayer::kSubAttrWindowRightMargin:
       case plusplayer::kSubAttrWindowTopMargin:
       case plusplayer::kSubAttrWindowBottomMargin:
-      case plusplayer::kSubAttrWindowBgColor:
       case plusplayer::kSubAttrFontWeight:
       case plusplayer::kSubAttrFontStyle:
-      case plusplayer::kSubAttrFontColor:
-      case plusplayer::kSubAttrFontBgColor:
-      case plusplayer::kSubAttrFontTextOutlineColor:
       case plusplayer::kSubAttrFontTextOutlineThickness:
       case plusplayer::kSubAttrFontTextOutlineBlurRadius:
       case plusplayer::kSubAttrFontVerticalAlign:
@@ -1190,35 +1268,111 @@ void PlusPlayer::OnSubtitleData(char *data, const int size,
       case plusplayer::kSubAttrWebvttCueAlign:
       case plusplayer::kSubAttrWebvttCuePositionAlign:
       case plusplayer::kSubAttrWebvttCueVertical:
+      case plusplayer::kSubAttrWindowShowBg:
       case plusplayer::kSubAttrTimestamp: {
-        int value_int = reinterpret_cast<int>(attr->value);
-        LOG_INFO("[PlusPlayer] Subtitle update: value<int>: %d", value_int);
+        const int *value_int = static_cast<const int *>(attr->value);
+        LOG_INFO("[PlusPlayer] Subtitle update: value<int>: %d", *value_int);
         attributes[flutter::EncodableValue("attrValue")] =
-            flutter::EncodableValue(value_int);
+            flutter::EncodableValue(*value_int);
       } break;
       case plusplayer::kSubAttrFontFamily:
       case plusplayer::kSubAttrRawSubtitle: {
-        const char *value_chars = reinterpret_cast<const char *>(attr->value);
+        const char *value_chars = static_cast<const char *>(attr->value);
         LOG_INFO("[PlusPlayer] Subtitle update: value<char *>: %s",
                  value_chars);
         std::string value_string(value_chars);
         attributes[flutter::EncodableValue("attrValue")] =
             flutter::EncodableValue(value_string);
       } break;
-      case plusplayer::kSubAttrWindowShowBg: {
-        uint32_t value_uint32 = reinterpret_cast<uint32_t>(attr->value);
+      case plusplayer::kSubAttrFontColor:
+      case plusplayer::kSubAttrFontBgColor:
+      case plusplayer::kSubAttrWindowBgColor:
+      case plusplayer::kSubAttrFontTextOutlineColor: {
+        const uint32_t *value_uint32 =
+            static_cast<const uint32_t *>(attr->value);
         LOG_INFO("[PlusPlayer] Subtitle update: value<uint32_t>: %u",
-                 value_uint32);
+                 *value_uint32);
         attributes[flutter::EncodableValue("attrValue")] =
-            flutter::EncodableValue((int64_t)value_uint32);
+            flutter::EncodableValue((int64_t)*value_uint32);
       } break;
       default:
-        LOG_ERROR("[PlusPlayer] Unknown Subtitle type: %d", attr->type);
+        LOG_INFO("[PlusPlayer] Unknown Subtitle type: %d", attr->type);
         break;
     }
-    attributes_list.push_back(flutter::EncodableValue(attributes));
+
+    int index = line_attr_index[attr->type]--;
+    if (index >= 0) {
+      attributes_lines[index].push_back(flutter::EncodableValue(attributes));
+    }
   }
-  self->SendSubtitleUpdate(duration, data, attributes_list);
+
+  if (type == plusplayer::SubtitleType::kPicture) {
+    if (picture_width <= 0 || picture_height <= 0 || size <= 0) {
+      LOG_ERROR(
+          "[PlusPlayer] Invalid picture dimensions or size: size: %d, width: "
+          "%f, height: %f",
+          size, picture_width, picture_height);
+      return;
+    }
+
+    const double area = picture_width * picture_height;
+    if (area > static_cast<double>(std::numeric_limits<int>::max()) ||
+        area > size) {
+      LOG_ERROR("[PlusPlayer] Picture area too large: %f", area);
+      return;
+    }
+
+    LOG_INFO(
+        "[PlusPlayer] Subtitle is a picture: size: %d, width: %f, height: %f",
+        size, picture_width, picture_height);
+
+    int subtitle_mem_length = 0;
+    int channels = size / area;
+    if (channels < 1 || channels > 4) {
+      LOG_ERROR("[PlusPlayer] Invalid number of channels: %d", channels);
+      return;
+    }
+    int stride_in_bytes = static_cast<int>(picture_width) * channels;
+
+    unsigned char *subtitle_png = stbi_write_png_to_mem(
+        (const unsigned char *)data, stride_in_bytes,
+        static_cast<int>(picture_width), static_cast<int>(picture_height),
+        channels, &subtitle_mem_length);
+
+    if (subtitle_png && subtitle_mem_length > 0) {
+      std::vector<uint8_t> picture(subtitle_png,
+                                   subtitle_png + subtitle_mem_length);
+      flutter::EncodableMap picture_info = {
+          {flutter::EncodableValue("picture"),
+           flutter::EncodableValue(picture)},
+          {flutter::EncodableValue("pictureWidth"),
+           flutter::EncodableValue(picture_width)},
+          {flutter::EncodableValue("pictureHeight"),
+           flutter::EncodableValue(picture_height)},
+      };
+      self->SendSubtitleUpdate(duration, flutter::EncodableList(),
+                               picture_info);
+      STBIW_FREE(subtitle_png);
+    } else {
+      LOG_ERROR("[PlusPlayer] Picture subtitle data is null or size is 0.");
+    }
+  } else {
+    LOG_INFO(
+        "[PlusPlayer] Subtitle is text: duration: %llu, text: %s, type: %d",
+        duration, data, type);
+    flutter::EncodableList texts_info = {};
+    for (int i = 0; i < text_lines_count; i++) {
+      flutter::EncodableMap text_line = {
+          {flutter::EncodableValue("text"),
+           flutter::EncodableValue(text_lines[i])},
+          {flutter::EncodableValue("attributes"),
+           flutter::EncodableValue(attributes_lines[i])},
+      };
+      texts_info.emplace_back(flutter::EncodableValue(text_line));
+    }
+
+    self->SendSubtitleUpdate(duration, texts_info);
+  }
 }
 
 void PlusPlayer::OnResourceConflicted(void *user_data) {
@@ -1321,8 +1475,8 @@ void PlusPlayer::OnDrmInitData(int *drm_handle, unsigned int len,
 void PlusPlayer::OnAdaptiveStreamingControlEvent(
     const plusplayer::StreamingMessageType &type,
     const plusplayer::MessageParam &msg, void *user_data) {
-  LOG_INFO("[PlusPlayer] Message type: %d, is DrmInitData (%d)", type,
-           type == plusplayer::StreamingMessageType::kDrmInitData);
+  LOG_DEBUG("[PlusPlayer] Message type: %d, is DrmInitData (%d)", type,
+            type == plusplayer::StreamingMessageType::kDrmInitData);
   PlusPlayer *self = reinterpret_cast<PlusPlayer *>(user_data);
 
   if (type == plusplayer::StreamingMessageType::kDrmInitData) {
@@ -1334,6 +1488,9 @@ void PlusPlayer::OnAdaptiveStreamingControlEvent(
     if (self->drm_manager_) {
       self->drm_manager_->UpdatePsshData(msg.data.data(), msg.size);
     }
+  }
+  if (type == plusplayer::StreamingMessageType::kManifestUpdated) {
+    self->SendManifestInfo(msg.data);
   }
 }