+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Adds compatibility with `http` 1.0 in example.
+* Removed unused proxy APIs.
+
 ## 0.5.8
 
 * Add namespace to the C++ code to avoid name conflicts with other plugins.